feat: Universal Container Image
This does a few things.
 * plain and keycloak containers are removed
 * universal container is update
 * README updated with universal container documentation
