+++ docker-compose/.env
@@ -8,3 +8,7 @@ KEYCLOAK_USER=admin
 KEYCLOAK_PASSWORD=admin
 KEYCLOAK_SERVICE_HOST=keycloak
 KEYCLOAK_SERVICE_PORT=8080
+ARTEMIS_USER=upsmessaging
+ARTEMIS_PASSWORD=upspassword
+ARTEMIS_SERVICE_HOST=artemis
+ARTEMIS_SERVICE_PORT=61617
