+++ jaxrs/pom.xml
@@ -183,6 +183,4 @@
             </dependencies>
         </profile>
     </profiles>
-
-
 </project>
