+++ push-sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/APNSClientProducer.java
@@ -7,7 +7,7 @@ import javax.ejb.Stateless;
 @Stateless
 public class APNSClientProducer extends AbstractJMSMessageProducer {
 
-    private static final String APNS_CLIENT = "topic/APNSClient";
+    private static final String APNS_CLIENT = "APNSClient";
 
     public void changeAPNClient(final iOSVariant iOSVariant) {
         super.sendTransacted(APNS_CLIENT, iOSVariant, true);
