--- servers/keycloak/src/main/docker/docker-compose.yml
@@ -1,20 +1,0 @@
-version: '2.1'
-
-services:
-  unifiedpushserver:
-    image: aerogear/ups:kc
-    volumes:
-       - ./helper:/ups-helper:z
-    entrypoint: "/ups-helper/exportKeycloakHost.sh"
-    links:
-      - keycloakServer:keycloak
-    ports:
-      - 18080:8080
-  keycloakServer:
-    image: docker.io/jboss/keycloak:3.4.3.Final
-    command: "-b 0.0.0.0 -Dkeycloak.import=/ups-keycloak-config/ups-realm-sample.json"
-    volumes:
-      - ./keycloak-realm:/ups-keycloak-config:z
-    environment:
-      KEYCLOAK_USER: admin
-      KEYCLOAK_PASSWORD: admin
