--- servers/keycloak/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,2 +1,0 @@
-<jboss-web>
-</jboss-web>
