--- servers/keycloak/src/main/webapp/WEB-INF/keycloak.json
@@ -1,9 +1,0 @@
-{
-  "realm" : "${ups.realm.name}",
-  "auth-server-url" : "${ups.auth.server.url}",
-  "ssl-required" : "external",
-  "resource" : "unified-push-server",
-  "bearer-only" : true,
-  "disable-trust-manager" : true
-}
-
