--- servers/keycloak/src/main/webapp/WEB-INF/web.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns="http://java.sun.com/xml/ns/javaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-         version="3.0">
-
-    <module-name>ag-push</module-name>
-
-    <security-constraint>
-        <web-resource-collection>
-            <web-resource-name>Configuration Endpoint</web-resource-name>
-            <url-pattern>/rest/auth/config</url-pattern>
-        </web-resource-collection>
-    </security-constraint>
-
-    <security-constraint>
-        <web-resource-collection>
-            <web-resource-name>Admin</web-resource-name>
-            <url-pattern>/rest/*</url-pattern>
-        </web-resource-collection>
-        <auth-constraint>
-            <role-name>admin</role-name>
-            <role-name>developer</role-name>
-        </auth-constraint>
-    </security-constraint>
-
-    <security-constraint>
-        <web-resource-collection>
-            <web-resource-name>HttpBasic</web-resource-name>
-            <description>Matches a few special URLs, not protected by Keycloak</description>
-            <url-pattern>/rest/registry/device/*</url-pattern>
-            <url-pattern>/rest/sender/*</url-pattern>
-            <url-pattern>/rest/prometheus/*</url-pattern>
-        </web-resource-collection>
-    </security-constraint>
-
-    <login-config>
-        <auth-method>KEYCLOAK</auth-method>
-        <realm-name>aerogear</realm-name>
-    </login-config>
-
-    <security-role>
-        <role-name>admin</role-name>
-    </security-role>
-    <security-role>
-        <role-name>developer</role-name>
-    </security-role>
-
-</web-app>
