--- servers/plain/pom.xml
@@ -1,172 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  JBoss, Home of Professional Open Source
-  Copyright Red Hat, Inc., and individual contributors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.jboss.aerogear.unifiedpush</groupId>
-        <artifactId>unifiedpush-parent</artifactId>
-        <version>2.2.2-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>unifiedpush-server-wildfly</artifactId>
-    <packaging>war</packaging>
-    <name>UnifiedPush Server for Wildfly (Plain)</name>
-
-    <properties>
-        <base>jboss/wildfly:16.0.0.Final</base>
-        <configFile>standalone-full.xml</configFile>
-        <postgresConfigFile>psql-module.xml</postgresConfigFile>    
-        <project.artifactId>${project.artifactId}</project.artifactId>
-    </properties>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.jboss.aerogear.unifiedpush</groupId>
-            <artifactId>unifiedpush-server-dependencies</artifactId>
-            <version>${project.version}</version>
-            <type>pom</type>
-        </dependency>
-
-        <dependency>
-            <groupId>org.aerogear.unifiedpush</groupId>
-            <artifactId>admin-ui</artifactId>
-            <version>${admin-ui.version}</version>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-core</artifactId>
-            <version>4.3.7.Final</version>
-            <scope>provided</scope>
-        </dependency>
-
-    </dependencies>
-
-    <repositories>
-        <repository>
-            <id>JBOSS_NEXUS</id>
-            <url>http://repository.jboss.org/nexus/content/groups/public</url>
-        </repository>
-    </repositories>
-
-    <build>
-        <resources>
-            <resource>
-                <!-- no need to include the database modules -->
-                <directory>src/main/resources/</directory>
-                <excludes>
-                    <exclude>modules/**/*.*</exclude>
-                </excludes>
-            </resource>
-        </resources>
-        <finalName>ag-push</finalName>
-        <plugins>
-            <plugin>
-                <artifactId>maven-war-plugin</artifactId>
-                <extensions>false</extensions>
-                <configuration>
-                    <failOnMissingWebXml>false</failOnMissingWebXml>
-                    <archive>
-                        <manifest>
-                            <addDefaultImplementationEntries>true</addDefaultImplementationEntries>
-                        </manifest>
-                        <manifestEntries>
-                            <Build-Branch>${scmBranch}</Build-Branch>
-                            <Implementation-Build>${buildNumber}</Implementation-Build>
-                            <Implementation-Build-Time>${maven.build.timestamp}</Implementation-Build-Time>
-                            <SCM-URL>${scm.url}</SCM-URL>
-                            <SCM-Tag>${scm.tag}</SCM-Tag>
-                        </manifestEntries>
-                    </archive>
-                </configuration>
-            </plugin>
-
-            <!-- To make the current git hash, we use the buildnumber-maven-plugin. -->
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>buildnumber-maven-plugin</artifactId>
-                <version>1.4</version>
-                <executions>
-                    <execution>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>create</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-    <profiles>
-        <profile>
-            <activation>
-                <property>
-                    <name>nodocker</name>
-                    <value>!true</value>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>io.fabric8</groupId>
-                        <artifactId>docker-maven-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>dmp</id>
-                                <phase>package</phase>
-                                <goals>
-                                    <goal>build</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                        <configuration>
-                            <images>
-                                <image>
-                                    <name>aerogear/unifiedpush-wildfly-plain:${version}</name>
-                                    <alias>server</alias>
-                                    <build>
-                                        <!-- filter>@</filter-->
-                                        <dockerFileDir>${project.basedir}/src/main/docker</dockerFileDir>
-                                        <assembly>
-                                            <descriptorRef>rootWar</descriptorRef>
-                                        </assembly>
-                                    </build>
-                                    <run>
-                                        <env>
-                                            <POSTGRES_SERVICE_HOST>${env.POSTGRES_SERVICE_HOST}</POSTGRES_SERVICE_HOST>
-                                            <POSTGRES_SERVICE_PORT>${env.POSTGRES_SERVICE_PORT}</POSTGRES_SERVICE_PORT>
-                                            <POSTGRES_DATABASE>${env.POSTGRES_DATABASE}</POSTGRES_DATABASE>
-                                            <POSTGRES_USER>${env.POSTGRES_USER}</POSTGRES_USER>
-                                            <POSTGRES_PASSWORD>${env.POSTGRES_PASSWORD}</POSTGRES_PASSWORD>
-                                        </env>
-                                        <ports>
-                                            <port>8080:8080</port>
-                                        </ports>
-                                    </run>
-                                </image>
-                            </images>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-</project>
