--- servers/plain/src/main/webapp/WEB-INF/jboss-deployment-structure.xml
@@ -1,9 +1,0 @@
-<jboss-deployment-structure>
-    <deployment>
-        <dependencies>
-            <module name="org.jboss.resteasy.resteasy-jackson2-provider"  services="import" />
-            <module name="org.jboss.xnio"/>
-        </dependencies>
-    </deployment>
-</jboss-deployment-structure>
-
