--- servers/plain/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,3 +1,0 @@
-<jboss-web>
-   <context-root>/</context-root>
-</jboss-web>
