+++ servers/universal/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,0 +1,2 @@
+<jboss-web>
+</jboss-web>
